requirements.txt
astroid==2.11.5
autopep8==1.6.0
awscrt==0.13.11
boto==2.49.0
boto3==1.24.2
botocore==1.27.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
dill==0.3.5.1
flake8==4.0.1
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
pendulum==2.1.2
platformdirs==2.5.2
pycodestyle==2.8.0
pyfiglet==0.8.post1
pyflakes==2.4.0
pylint==2.14.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytzdata==2020.1
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
